Source maps - show original filename.
If the preprocessor renames files (eg. foo.es6.js -> foo.js).

Also, store the source map on file object, so that we can rewrite stace traces. When there is an exception in a compiled file (with source map), Chrome web inspector shows the original line numbers, but it is a feature of Chrome, not V8. Therefore the actual exception contains the compiled numbers. In order to show the original line numbers in Karma (commandline), we need to do the rewriting on our own. That's why I'm storing the source map, so that a reporter can read it and override exceptions.